Restructured source files to use src directory
